EP 05 BACKEND FUNCTIONALITY FINISHED

+++ TodoListApp.WebApi/Controllers/TagController.cs
@@ -1,5 +1,7 @@
 using Microsoft.AspNetCore.Mvc;
+using TodoListApp.Services.Exceptions;
 using TodoListApp.Services.Interfaces;
+using TodoListApp.WebApi.Logging;
 using TodoListApp.WebApi.Models.Tags;
 
 namespace TodoListApp.WebApi.Controllers;
@@ -8,17 +10,42 @@ namespace TodoListApp.WebApi.Controllers;
 public class TagController : ControllerBase
 {
     private readonly ITagService tagService;
+    private readonly ILogger<TagController> logger;
 
-    public TagController(ITagService tagService)
+    public TagController(ITagService tagService, ILogger<TagController> logger)
     {
         this.tagService = tagService;
+        this.logger = logger;
     }
 
     [HttpGet]
     public async Task<ActionResult<IEnumerable<TagDto>>> GetAllTags()
     {
-        var tags = await this.tagService.GetAllTagsAsync();
-        return this.Ok(tags);
+        try
+        {
+            var tags = await this.tagService.GetAllTagsAsync();
+
+            if (!tags.Any())
+            {
+                TagControllerLoggerMessages.TagNotFoundWhileGettingAllTags(this.logger);
+            }
+            else
+            {
+                TagControllerLoggerMessages.TagsRetrieved(this.logger);
+            }
+
+            return this.Ok(tags);
+        }
+        catch (InvalidOperationException ioe)
+        {
+            TagControllerLoggerMessages.UnexpectedErrorOccurredWhileGettingAllTags(this.logger, ioe.Message, ioe);
+            return this.StatusCode(StatusCodes.Status500InternalServerError, "An invalid operation occured" + ioe.Message);
+        }
+        catch (Exception ex)
+        {
+            TagControllerLoggerMessages.UnexpectedErrorOccurredWhileGettingAllTags(this.logger, ex.Message, ex);
+            throw;
+        }
     }
 
     [HttpPost("AddTagToTask")]
@@ -26,30 +53,66 @@ public class TagController : ControllerBase
     {
         if (string.IsNullOrWhiteSpace(tagName))
         {
+            TagControllerLoggerMessages.AddTagToTaskLogWarningTagEmpty(this.logger);
             return this.BadRequest("Tag name cannot be empty.");
         }
 
         try
         {
+            bool isTagNameUnique = await this.tagService.IsTagNameUniqueAsync(tagName);
+            if (!isTagNameUnique)
+            {
+                TagControllerLoggerMessages.AddTagToTaskLogWarningTagExists(this.logger, tagName);
+                return this.BadRequest($"A tag with the name '{tagName}' already exists.");
+            }
+
             var tagDto = await this.tagService.AddTagToTaskAsync(tagName, taskId);
+            TagControllerLoggerMessages.TagAddedToTask(this.logger, tagName, taskId);
             return this.Ok(tagDto);
         }
-        catch (InvalidOperationException ioe)
+        catch (ServiceException se)
         {
-            return this.StatusCode(StatusCodes.Status500InternalServerError, "An invalid operation occured" + ioe.Message);
+            TagControllerLoggerMessages.ServiceExceptionOccurredWhileAddingTag(this.logger, se.Message, se);
+            return this.StatusCode(StatusCodes.Status500InternalServerError, se.Message);
+        }
+        catch (Exception ex)
+        {
+            TagControllerLoggerMessages.UnexpectedErrorOccurredWhileAddingTag(this.logger, ex.Message, ex);
+            throw;
         }
     }
 
     [HttpDelete("{taskId}/tags/{tagId}")]
     public async Task<IActionResult> DeleteTag(int taskId, int tagId)
     {
-        var result = await this.tagService.DeleteTagAsync(taskId, tagId);
-
-        if (result)
+        try
         {
-            return this.NoContent();
-        }
+            if (taskId <= 0 || tagId <= 0)
+            {
+                return this.BadRequest("Invalid taskId or tagId.");
+            }
 
-        return this.NotFound();
+            var result = await this.tagService.DeleteTagAsync(taskId, tagId);
+
+            if (result)
+            {
+                return this.NoContent();
+            }
+
+            return this.NotFound();
+        }
+        catch (KeyNotFoundException)
+        {
+            return this.NotFound();
+        }
+        catch (ServiceException se)
+        {
+            return this.StatusCode(StatusCodes.Status500InternalServerError, se.Message);
+        }
+        catch (Exception ex)
+        {
+            TagControllerLoggerMessages.UnexpectedErrorOccurredWhileDeletingTag(this.logger, ex.Message, ex);
+            throw;
+        }
     }
 }
